pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baserun"
version = "0.9.8b1"
description = "Tools for testing, debugging, and evaluating LLM features."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Adam Ginzberg", email = "adam@baserun.ai" },
    { name = "Erik Peterson", email = "erik@baserun.ai" }
]
requires-python = ">=3.7.1"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.1.1",
    "wrapt~=1.15.0",
    "opentelemetry-instrumentation>=0.43b0",
    "opentelemetry-sdk>=1.20.0",
    "grpcio~=1.58.0",
    "protobuf~=4.24.3",
    "grpcio-tools~=1.58.0",
    "python-Levenshtein~=0.23.0"
]

[project.urls]
Homepage = "https://baserun.ai"
Repository = "https://github.com/baserun-ai/baserun-py"

[project.entry-points.pytest11]
pbaserun = "baserun.pytest_plugin"